split config.updateFromArgv() into getArgv() and config.updateFromArgv(argv), add constants.js (#333)
* split config.updateFromArgv() into getArgv() and config.updateFromArgv(argv)

also add constants.js

* move droppedOptions to constants.js

* Simplified exports

* add newline at end of file
